+++ src/main/controllers/fs.ts
@@ -4,21 +4,25 @@
  */
 
 import { Controller } from '@/controllers/controller';
-import { IpcFsReadDirArgs, ipcFsReadDirChannel, IpcFsReadDirRes, IpcFsGetHomeDirArgs, ipcFsGetHomeDirChannel, IpcFsGetHomeDirRes } from '@common/ipc/channels';
+import { IpcFsReadDirArgs, ipcFsReadDirChannel, IpcFsReadDirRes, IpcFsGetHomeDirArgs, ipcFsGetHomeDirChannel, IpcFsGetHomeDirRes, IpcFsGetImageDataUrlArgs, ipcFsGetImageDataUrlChannel, IpcFsGetImageDataUrlRes } from '@common/ipc/channels';
 import { ReadDirUseCase } from '@/application/useCases/fs/readDir';
 import { GetHomeDirUseCase } from '@/application/useCases/fs/getHomeDir';
+import { GetImageDataUrlUseCase } from '@/application/useCases/fs/getImageDataUrl';
 
 type Deps = {
   readDirUseCase: ReadDirUseCase;
   getHomeDirUseCase: GetHomeDirUseCase;
+  getImageDataUrlUseCase: GetImageDataUrlUseCase;
 }
 
 export function createFsControllers({
   readDirUseCase,
   getHomeDirUseCase,
+  getImageDataUrlUseCase,
 }: Deps): [
     Controller<IpcFsReadDirArgs, IpcFsReadDirRes>,
     Controller<IpcFsGetHomeDirArgs, IpcFsGetHomeDirRes>,
+    Controller<IpcFsGetImageDataUrlArgs, IpcFsGetImageDataUrlRes>,
   ] {
   return [{
     channel: ipcFsReadDirChannel,
@@ -26,5 +30,8 @@ export function createFsControllers({
   }, {
     channel: ipcFsGetHomeDirChannel,
     handle: async () => getHomeDirUseCase()
+  }, {
+    channel: ipcFsGetImageDataUrlChannel,
+    handle: async (_event, path) => getImageDataUrlUseCase(path)
   }]
 }