+++ src/renderer/infra/fsProvider/fsProvider.ts
@@ -3,7 +3,7 @@
  * GNU General Public License v3.0 or later (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
  */
 
-import { IpcFsReadDirArgs, ipcFsReadDirChannel, IpcFsReadDirRes, IpcFsGetHomeDirArgs, ipcFsGetHomeDirChannel, IpcFsGetHomeDirRes } from '@common/ipc/channels';
+import { IpcFsReadDirArgs, ipcFsReadDirChannel, IpcFsReadDirRes, IpcFsGetHomeDirArgs, ipcFsGetHomeDirChannel, IpcFsGetHomeDirRes, IpcFsGetImageDataUrlArgs, ipcFsGetImageDataUrlChannel, IpcFsGetImageDataUrlRes } from '@common/ipc/channels';
 import { electronIpcRenderer } from '@/infra/mainApi/mainApi';
 import { FsProvider } from '@/application/interfaces/fsProvider';
 
@@ -16,6 +16,10 @@ export function createFsProvider(): FsProvider {
     ),
     getHomeDir: () => electronIpcRenderer.invoke<IpcFsGetHomeDirArgs, IpcFsGetHomeDirRes>(
       ipcFsGetHomeDirChannel
+    ),
+    getImageDataUrl: (path) => electronIpcRenderer.invoke<IpcFsGetImageDataUrlArgs, IpcFsGetImageDataUrlRes>(
+      ipcFsGetImageDataUrlChannel,
+      path
     ),
   }
 }