+++ tests/main/infra/mocks/fsProvider.ts
@@ -7,7 +7,8 @@ import { FsProvider } from '@/application/interfaces/fsProvider';
 
 const fsProvider: FsProvider = {
   readDir: jest.fn(async () => []),
-  getHomeDir: jest.fn(() => '')
+  getHomeDir: jest.fn(() => ''),
+  getImageDataUrl: jest.fn(async () => null)
 }
 
 export const mockFsProvider = (props: Partial<FsProvider>) => ({ ...fsProvider, ...props });